refactor(uec): refactor fully the UEC module for clarity and better logical structure
