+++ packages/devqubit-engine/src/devqubit_engine/uec/models/result.py
@@ -22,7 +22,7 @@ import traceback
 from dataclasses import dataclass, field
 from typing import Any
 
-from devqubit_engine.uec.types import ArtifactRef
+from devqubit_engine.uec.models.types import ArtifactRef
 
 
 @dataclass
